Extra args to process using existing file and start at specific uuid.
git-svn-id: http://ala-portal.googlecode.com/svn/trunk/biocache-store@2479 1ef1b718-b4f5-11dd-90a5-ed332e521da9

+++ src/main/scala/au/org/ala/util/ReprocessIndexSelect.scala
@@ -11,29 +11,35 @@ object ReprocessIndexSelect {
     def main(args: Array[String]): Unit = {
         var query:Option[String]=None;
         var threads = 4
+        var exist = false;
+        var startUuid:Option[String]=None
         val parser = new OptionParser("index records options") {
             opt("q", "query", "The query to run", { v:String => query = Some(v) })
             intOpt("t", "thread", "The number of threads to use", {v:Int => threads = v } )
+            opt("exist", "use the existing list of row keys", {exist=true})
+            opt("s", "start","The record to start processing with", {v:String => startUuid = Some(v)})
             
         }
 
         if (parser.parse(args)) {
             if(!query.isEmpty){
-                reprocessindex(query.get, threads)
+                reprocessindex(query.get, threads, exist, startUuid)
             }
         }
     }
 
-    def reprocessindex(query: String, threads: Int) {
+    def reprocessindex(query: String, threads: Int, exist: Boolean, startUuid:Option[String]) {
         //get the list of rowKeys to be processed.
         val indexer = Config.getInstance(classOf[IndexDAO]).asInstanceOf[IndexDAO]
 
         val file = new File("rowkeys.out")
+        if(!exist){
         val out = new BufferedOutputStream(new FileOutputStream(file));        
-        indexer.writeRowKeysToStream(query,out)
-        out.flush()
-        out.close()
-        ProcessWithActors.processRecords(threads, file)
+            indexer.writeRowKeysToStream(query,out)
+            out.flush()
+            out.close()
+        }
+        ProcessWithActors.processRecords(threads, file, startUuid)
         IndexRecords.indexList(file)
 
     }
